+++ solvers/ADRSolver/EquationSystems/SteadyAdvectionDiffusion.h
@@ -31,6 +31,7 @@
 // Description: Steady advection-diffusion solve routines
 //
 ///////////////////////////////////////////////////////////////////////////////
+
 #ifndef NEKTAR_SOLVERS_ADRSOLVER_EQUATIONSYSTEMS_ADVECTIONDIFFUSION_H
 #define NEKTAR_SOLVERS_ADRSOLVER_EQUATIONSYSTEMS_ADVECTIONDIFFUSION_H