+++ solvers/ADRSolver/EquationSystems/SteadyAdvectionDiffusionReaction.cpp
@@ -34,11 +34,9 @@
 
 #include <ADRSolver/EquationSystems/SteadyAdvectionDiffusionReaction.h>
 
-using namespace std;
-
 namespace Nektar
 {
-string SteadyAdvectionDiffusionReaction::className =
+std::string SteadyAdvectionDiffusionReaction::className =
     GetEquationSystemFactory().RegisterCreatorFunction(
         "SteadyAdvectionDiffusionReaction",
         SteadyAdvectionDiffusionReaction::create);